refactoring and subnav design
