+++ content/en-US/ROADMAP.md
@@ -1,0 +1,41 @@
+## 2017
+
+### October
+* Alpha testnet for the Cosmos Hub (just staking, no ibc or governance)
+* Translation of Whitepaper into Chinese
+
+### November
+* Public testnet for the Cosmos Hub
+* Alpha testnet for Ethermint
+* Beta version of the Cosmos SDK
+* Translation of whitepaper into Dutch and French
+
+### December
+* Public testnet for Ethermint
+* Alpha Ethereum peg zone
+* Stable version of the Cosmos SDK
+* Translation of Whitepaper into Spanish and Korean
+
+## 2018
+
+### January
+* Live Cosmos Hub
+* Public testnet Ethereum peg zone
+* IBC connection between Ethermint and Cosmos Hub on testnets
+
+### February
+* Live Ethermint network
+* Connect Ethermint to Cosmos Hub with IBC
+
+### March
+* First token launches are conducted on Ethermint to create new zones
+
+### April
+* Bitcoin peg zone
+* Incubator zone to bootstrap projects that do not have the resources to launch their own validator set
+
+### May
+* Work on arbitrary data transfer
+
+### June
+* Identity zone launches on Cosmos to enable anonymous and decentralised KYC
