+++ src/components/NavFaq.vue
