+++ src/components/NavWhitepaper.vue
