+++ src/components/pages/PageDev.vue
@@ -1,0 +1,21 @@
+<template lang="pug">
+.page
+  page-menu
+    router-link(to="/dev") Overview
+    router-link(to="/dev/whitepaper") Whitepaper
+    router-link(to="/dev/hackatom") HackAtom
+  router-view
+</template>
+
+<script>
+import PageMenu from 'common/NiPageMenu'
+export default {
+  name: 'page-intro',
+  components: {
+    PageMenu
+  },
+  mounted () {
+    document.title = 'Developers - Cosmos - Internet of Blockchains'
+  }
+}
+</script>
