+++ src/components/pages/PageIndex.vue
@@ -31,7 +31,7 @@
     </section>
     <section class="section-home">
       <div class="section-container">
-        <div class="image"><img class="" src="../assets/images/ex_dist_exchange.png"></div>
+        <div class="image"><img class="" src="~assets/images/ex_dist_exchange.png"></div>
         <div class="section-content">
           <h2>Interoperability</h2>
           <p>Cosmos can interoperate with multiple other applications and cryptocurrencies, something other blockchains can’t do well. By creating a new zone, you can plug any blockchain system into the Cosmos hub and pass tokens back and forth between those zones, without the need for an intermediary.<br><br> For example, if you wanted to trade bitcoins for ether today, you would have to do so on a centralized exchange. The problem is, many of our exchanges today are plagued by hacks, theft, and cons. Two prime examples are what happened at <a href='http://www.thedailybeast.com/articles/2016/05/19/behind-the-biggest-bitcoin-heist-in-history-inside-the-implosion-of-mt-gox.html'>Mt. Gox</a> and more recently <a href='http://www.reuters.com/article/us-bitfinex-hacked-hongkong-idUSKCN10E0KP'>Bitfinex</a>.</p>
@@ -40,7 +40,7 @@
     </section>
     <section class="section-home sh-alt">
       <div class="section-container">
-        <div class="image"><img class="" src="../assets/images/ex_evm_sharding.png"></div>
+        <div class="image"><img class="" src="~assets/images/ex_evm_sharding.png"></div>
         <div class="section-content">
           <h2>Scalability</h2>
           <p>Scaling is another open issue for blockchains. Both Ethereum and Bitcoin support only a fraction of the transactions seen daily on payment networks like Visa or Mastercard. In contrast, zones allow Cosmos to scale out indefinitely. If your transaction speed slows in one zone because too many people are using it, you simply add another zone to the hub and direct half the users over that zone, thereby doubling your transaction speed. Meanwhile, the Cosmos hub ensures that any zones connected to it remain in sync.</p>
@@ -49,7 +49,7 @@
     </section>
     <section class="section-home">
       <div class="section-container">
-        <div class="image"><img class="" src="../assets/images/ex_evm_upgrading.png"></div>
+        <div class="image"><img class="" src="~assets/images/ex_evm_upgrading.png"></div>
         <div class="section-content">
           <h2>Upgradability</h2>
           <p>Another problem blockchains run into is how to handle upgrades when new versions come out. Getting all of the validators ('miners' in the case of Bitcoin) to upgrade to a new version of a blockchain simultaneously is tricky, and could lead to hard forks. This is exactly what happened when <a href='http://qz.com/730004/everything-you-need-to-know-about-the-ethereum-hard-fork/'>Ethereum hard-forked</a> into ETH and ETC due to political/governance issues surrounding the DAO hack. With Cosmos, upgrading is not an issue. You simply plug a new zone into the Cosmos hub, and invite users to move their funds over to it at their leisure.</p>
@@ -58,7 +58,7 @@
     </section>
     <section class="section-home sh-alt">
       <div class="section-container">
-        <div class="image image-taller"><img class="" src="../assets/images/ex_network.png"></div>
+        <div class="image image-taller"><img class="" src="~assets/images/ex_network.png"></div>
         <div class="section-content">
           <h2>Decentralization</h2>
           <p>One of the great tragedies of cryptocurrencies today has been their failure to interoperate with existing systems — and each other. Cosmos offers a way to link them together, but without creating a new centralized clearinghouse or exchange.<br><br>Cosmos is built on principles of cryptography, sound economics, consensus theory, transparency, and accountability to serve as a new foundation for our future financial systems.</p>
@@ -74,7 +74,7 @@
 
 <script>
 import {mapGetters} from 'vuex'
-import SectionCover from './SectionCover'
+import SectionCover from 'comp/SectionCover'
 import Btn from '@nylira/vue-button'
 export default {
   name: 'page-index',
@@ -101,7 +101,7 @@ export default {
 </script>
 
 <style lang="stylus">
-@import '../styles/variables.styl'
+@import '~variables'
 
 .section-home
   &#home-intro
