+++ src/components/pages/PageWhitepaper.vue
@@ -1,9 +1,8 @@
 <template lang="pug">
-// toc-page(:toc-visible='whitepaperTocVisible')
-  //page-nav(slot='nav')
-part(title=" ")
-  div(slot="menu"): a(@click.native="download()") Download PDF
+// toc-page(:toc-visible='whitepaperTocVisible'): page-nav(slot='nav')
 
+page(title="Whitepaper" subtitle="Read the original Cosmos whitepaper.")
+  div(slot="menu"): btn(@click.native="download()" icon="print" value="Download PDF")
   text-container
     text-ko(v-if="$route.params.locale === 'ko'")
     text-pt(v-else-if="$route.params.locale === 'pt'")
@@ -13,19 +12,23 @@ part(title=" ")
 
 <script>
 import { mapGetters } from 'vuex'
-import TextContainer from './common/NiTextContainer'
-import Part from './common/NiPart'
-import PageNav from './PageWhitepaperNav'
-import TextEnUs from '../../content/en-US/WHITEPAPER.md'
-import TextKo from '../../content/ko/WHITEPAPER.md'
-import TextPt from '../../content/pt/WHITEPAPER.md'
-import TextZhCn from '../../content/zh-CN/WHITEPAPER.md'
+import Btn from '@nylira/vue-button'
+import Page from 'common/NiPage'
+import Part from 'common/NiPart'
+import PageNav from 'comp/NavWhitepaper'
+import TextContainer from 'common/NiTextContainer'
+import TextEnUs from 'content/en-US/WHITEPAPER.md'
+import TextKo from 'content/ko/WHITEPAPER.md'
+import TextPt from 'content/pt/WHITEPAPER.md'
+import TextZhCn from 'content/zh-CN/WHITEPAPER.md'
 // import TocPage from './TocPage'
 export default {
   components: {
-    TextContainer,
+    Btn,
     PageNav,
+    Page,
     Part,
+    TextContainer,
     TextEnUs,
     TextKo,
     TextPt,
