+++ src/styles/app.styl
@@ -42,13 +42,13 @@ h1, .h1, h2, .h2, h3, .h3, h4, .h4, h5, .h5, h6, .h6
     margin-bottom 0
 
 h1, .h1
-  font-size h1 * 0.7
+  font-size h1 * 0.875
 
 h2, .h2
-  font-size h2 * 0.8
+  font-size h2 * 0.875
 
 h3, .h3
-  font-size h3 * 0.9
+  font-size h3
 
 h4, .h4
   font-size h4
